+++ packages/docs-markdown/src/extension.ts
@@ -30,7 +30,7 @@ import { codeFormattingCommand } from './controllers/code-controller';
 import { insertImageCommand } from './controllers/image-controller';
 import { insertIncludeCommand } from './controllers/include-controller';
 import { italicFormattingCommand } from './controllers/italic-controller';
-import { linkControllerCommands } from './controllers/link-controller';
+import { linkControllerCommands } from './controllers/links/link-controller';
 import { insertListsCommands } from './controllers/list-controller';
 import { insertLinksAndMediaCommands } from './controllers/media-controller';
 import { insertMetadataCommands } from './controllers/metadata-controller';
