+++ packages/docs-markdown/src/test/suite/controllers/quick-pick-menu-controller.test.ts
@@ -19,7 +19,7 @@ import * as alertController from '../../../controllers/alert-controller';
 import * as listController from '../../../controllers/list-controller';
 import * as tableController from '../../../controllers/table-controller';
 import * as rowColumnsController from '../../../controllers/row-columns-controller';
-import * as linkController from '../../../controllers/link-controller';
+import * as linkController from '../../../controllers/links/link-controller';
 import * as noLocController from '../../../controllers/no-loc-controller';
 import * as imageController from '../../../controllers/image-controller';
 import * as includeController from '../../../controllers/include-controller';
